Init using sync.Once when targeting tinygo
